Backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
newsapi-python>=0.2.7
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
